+++ test/test_ul4.py
@@ -4686,6 +4686,8 @@ def test_method_insert(T):
 def test_method_pop(T):
 	assert '42;17;23;' == T('<?code l = [17, 23, 42]?><?print l.pop()?>;<?print l.pop(-2)?>;<?print l.pop(0)?>;').renders()
 	assert '42;17;23;' == T('<?code l = [17, 23, 42]?><?code m = l.pop?><?print m()?>;<?print m(-2)?>;<?print m(0)?>;').renders()
+	assert '23;73;{}' == T('<?code d = {17: 23, 42: 73}?><?print d.pop(17)?>;<?print d.pop(42)?>;<?print d?>').renders()
+	assert '23;42;{42: 73}' == T('<?code d = {17: 23, 42: 73}?><?print d.pop(17)?>;<?print d.pop(43, 42)?>;<?print d?>').renders()
 
 
 @pytest.mark.ul4